Moved appsettings to utilities
