+++ FolderInspector/FolderInspector.csproj
@@ -71,12 +71,13 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Constants\AppSettings.cs" />
     <Compile Include="FolderInspector.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Utilities\AppSettingsUtility.cs" />
     <Compile Include="Utilities\ConsoleLogUtility.cs" />
     <Compile Include="Utilities\ExcelUtility.cs" />
     <Compile Include="Utilities\FolderUtility.cs" />
+    <Compile Include="Utilities\IAppSettingsUtility.cs" />
     <Compile Include="Utilities\IDocumentUtility.cs" />
     <Compile Include="Utilities\ILogUtility.cs" />
     <Compile Include="Utilities\WordUtility.cs" />
@@ -89,6 +90,7 @@
     </Content>
   </ItemGroup>
   <ItemGroup>
+    <Folder Include="Constants\" />
     <Folder Include="Model\" />
   </ItemGroup>
   <ItemGroup>
